Moved some user profile related classes/files
