+++ ProxiPalTest_client_20231205232005/app/src/main/java/com/mongodb/app/presentation/userprofiles/UserProfileContextualMenuViewModel.kt
@@ -1,16 +1,11 @@
-package com.mongodb.app.presentation.tasks
+package com.mongodb.app.presentation.userprofiles
 // TODO Might need to move this class to a new package to avoid confusion being in the "tasks" package
 
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.State
-import androidx.compose.runtime.mutableStateOf
-import androidx.lifecycle.ViewModel
 import com.mongodb.app.data.SyncRepository
 import com.mongodb.app.domain.UserProfile
+import com.mongodb.app.presentation.tasks.ContextualMenuViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
 
 class UserProfileContextualMenuViewModel constructor(
